+++ narrenschiff/modules/gcloud.py
@@ -4,8 +4,6 @@ from narrenschiff.modules.common import NarrenschiffModule
 class Gcloud(NarrenschiffModule):
     """``gcloud`` module."""
 
-    gcloud_cmd = 'gcloud'
-
     @property
     def cmd(self):
         command = self.command.get('command')
@@ -18,4 +16,4 @@ class Gcloud(NarrenschiffModule):
 
         flags.extend(['--{}'.format(switch) for switch in switches])
 
-        return ' '.join([Gcloud.gcloud_cmd, command, *flags])
+        return ' '.join(['gcloud', command, *flags])